move ::tokenize() to lexer module, move ::Lexer out of header
