+++ src/interpreter/execute.cpp
@@ -5,6 +5,7 @@
 #include "../compute_offsets.hpp"
 #include "../cst_allocator.hpp"
 #include "../ct_eval.hpp"
+#include "../lexer.hpp"
 #include "../match_identifiers.hpp"
 #include "../metacheck.hpp"
 #include "../parser.hpp"
