+++ src/parser.hpp
@@ -20,6 +20,5 @@ struct Writer {
 	}
 };
 
-TokenArray tokenize(std::string const& source);
 Writer<CST::CST*> parse_program(TokenArray&, CST::Allocator&);
 Writer<CST::CST*> parse_expression(TokenArray&, CST::Allocator&);
